[Interpreter] Fix missing entries for bytecode handlers in perf mapping.
Since [1], Logger does not dispatch anymore to the other loggers. This is
now the task of the CodeEventDispatcher, and Logger is just a client.

One use of the Logger as dispatcher wasn't caught in the refactoring, which
caused the entries for bytecode handlers loaded from the snapshot to be
missing from the symbols mapping written by PerfBasicLogger.

[1] https://crrev.com/9df23510eaa12eb5c298cf68a92e90c6e992855c

BUG=v8:4899
LOG=N

Review-Url: https://codereview.chromium.org/2085143003
Cr-Commit-Position: refs/heads/master@{#37224}